src/metrics.py
# Metrics for Code-Switching Analysis
# Functions to compute linguistic metrics like marker retention, similarity scores, etc.

import re
from typing import List, Dict, Any
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity

def compute_marker_retention(original_text: str, generated_text: str, markers: List[str]) -> float:
    """
    Compute the percentage of linguistic markers retained in generated text.
    
    Args:
        original_text: Original text with code-switching markers
        generated_text: Model-generated text
        markers: List of linguistic markers to track
    
    Returns:
        Retention percentage (0-1)
    """
    # Placeholder implementation
    pass

def compute_semantic_similarity(text1: str, text2: str) -> float:
    """
    Compute semantic similarity between two texts using embeddings.
    
    Args:
        text1: First text
        text2: Second text
    
    Returns:
        Similarity score (0-1)
    """
    # Placeholder implementation
    pass

def analyze_code_switching_patterns(text: str) -> Dict[str, Any]:
    """
    Analyze code-switching patterns in text.
    
    Args:
        text: Input text to analyze
    
    Returns:
        Dictionary with pattern analysis results
    """
    # Placeholder implementation
    pass