requirements.txt
# Core dependencies for code-switching analysis
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# API clients
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
cohere>=4.0.0

# Data processing
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.64.0
pathlib>=1.0.0

# Text processing
nltk>=3.8.0
spacy>=3.4.0
transformers>=4.20.0
torch>=1.12.0